+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/TxInfo.hs
@@ -33,7 +33,12 @@ import Cardano.Ledger.BaseTypes (Inject (..), ProtVer (..), strictMaybe)
 import Cardano.Ledger.Coin (Coin (..))
 import Cardano.Ledger.Conway.Scripts (ConwayPlutusPurpose (..))
 import Cardano.Ledger.Conway.TxCert (Delegatee (..))
-import Cardano.Ledger.Conway.TxInfo (ConwayContextError (..), ConwayEraPlutusTxInfo (..))
+import Cardano.Ledger.Conway.TxInfo (
+  ConwayContextError (..),
+  ConwayEraPlutusTxInfo (..),
+  transTxInInfoV1,
+  transTxInInfoV3,
+ )
 import qualified Cardano.Ledger.Conway.TxInfo as Conway
 import Cardano.Ledger.Dijkstra.Core
 import Cardano.Ledger.Dijkstra.Era (DijkstraEra)
@@ -175,6 +180,8 @@ instance EraPlutusTxInfo 'PlutusV1 DijkstraEra where
 
   toPlutusArgs = Alonzo.toPlutusV1Args
 
+  toPlutusTxInInfo _ = transTxInInfoV1
+
 transTxCertV1V2 ::
   ( ConwayEraTxCert era
   , Inject (ConwayContextError era) (ContextError era)
@@ -235,6 +242,8 @@ instance EraPlutusTxInfo 'PlutusV2 DijkstraEra where
 
   toPlutusArgs = Babbage.toPlutusV2Args
 
+  toPlutusTxInInfo _ = Babbage.transTxInInfoV2
+
 instance EraPlutusTxInfo 'PlutusV3 DijkstraEra where
   toPlutusTxCert _ _ = pure . transTxCert
 
@@ -288,6 +297,8 @@ instance EraPlutusTxInfo 'PlutusV3 DijkstraEra where
       txBody = ltiTx ^. bodyTxL
 
   toPlutusArgs = Conway.toPlutusV3Args
+
+  toPlutusTxInInfo _ = transTxInInfoV3
 
 transTxCert ::
   (ConwayEraTxCert era, TxCert era ~ DijkstraTxCert era) => TxCert era -> PV3.TxCert
@@ -382,6 +393,8 @@ instance EraPlutusTxInfo 'PlutusV4 DijkstraEra where
       txBody = ltiTx ^. bodyTxL
 
   toPlutusArgs = toPlutusV4Args
+
+  toPlutusTxInInfo _ = transTxInInfoV3
 
 toPlutusV4Args ::
   EraPlutusTxInfo 'PlutusV4 era =>